Finish policies for controllers. Change foreign key for people

+++ app/Http/Controllers/Api/CustomerPersonController.php
@@ -14,9 +14,11 @@ class CustomerPersonController extends Controller
      * Display a listing of the resource.
      *
      * @return \Illuminate\Http\Response
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function index(Request $request)
     {
+        $this->authorize('viewAny');
         $posts = CustomerPerson::latest()->paginate(25);
 
         return $posts;
@@ -28,10 +30,11 @@ class CustomerPersonController extends Controller
      * @param \Illuminate\Http\Request $request
      *
      * @return \Illuminate\Http\Response
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function store(Request $request)
     {
-        
+        $this->authorize('create');
         $post = CustomerPerson::create($request->all());
 
         return response()->json($post, 201);
@@ -40,12 +43,14 @@ class CustomerPersonController extends Controller
     /**
      * Display the specified resource.
      *
-     * @param  int  $id
+     * @param int $id
      *
      * @return \Illuminate\Http\Response
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function show($id)
     {
+        $this->authorize('view');
         $post = CustomerPerson::findOrFail($id);
 
         return $post;
@@ -55,13 +60,14 @@ class CustomerPersonController extends Controller
      * Update the specified resource in storage.
      *
      * @param \Illuminate\Http\Request $request
-     * @param  int  $id
+     * @param int $id
      *
      * @return \Illuminate\Http\Response
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function update(Request $request, $id)
     {
-        
+        $this->authorize('edit');
         $post = CustomerPerson::findOrFail($id);
         $post->update($request->all());
 
@@ -71,12 +77,14 @@ class CustomerPersonController extends Controller
     /**
      * Remove the specified resource from storage.
      *
-     * @param  int  $id
+     * @param int $id
      *
      * @return \Illuminate\Http\Response
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function destroy($id)
     {
+        $this->authorize('delete');
         CustomerPerson::destroy($id);
 
         return response()->json(null, 204);
